SideBar: Device item click to mount, unmount widget
Device item in sidebar is now clickable and will mount the device if
possible. When mounted, an unmount widget will display in the right side
of the item and by clicking that item, user can unmount the device.
Notice that the position problem of context widget will be fixed in
another commit(which may change the contextWidget attach logic).

Change-Id: Ice7274563687b8779ea8f11ab246c3c7b6894efc