Fix the crash in getMountPointOfFile function
Previously, the function uses an iterator `rootPathPartionAndMountpointItr` to store
the partition and mount point of the root path. The iterator becomes
invalid when the iteration is over, therefore it causes SEGMENT FAULT in
some cases. In this fix, the iterator is replaced with a pair
`rootPathPartionAndMountpointItr` to store the information.

Resolves: https://github.com/linuxdeepin/dde-file-manager/issues/43