fix: use Q_GLOBAL_STATIC to initialise `urlToFileInfoMap`
During library unload or program exit, destruction of global objects is expected
to run before destruction of `urlToFileInfoMap`. But that's not guaranteed by
C++ standard and the program will segfault at exit. Instead, use Q_GLOBAL_STATIC
to initialize `urlToFileInfoMap`. So we can skip the cleanup `urlToFileInfoMap`
if it is already destructed.

Log: use Q_GLOBAL_STATIC to initialise `urlToFileInfoMap`
Signed-off-by: Robin Lee <cheeselee@fedoraproject.org>