fix: DFileDialog still show a removed stash mount
when remove a stash mount via file-manager's context menu, the `reload` signal is emitted in appcontroller `actionRemoveStashMounts`, but in dfiledialog, this signal will never be emitted, because DFileDialog will never trigger this function.
so I changed the place where the signal emitted. when the config file is changed, and the changed key is `RemoteMounts`, then emit the signal.
and remove some signal emit, no longer needed.

Log: 修复从文管移除常驻挂载后,文件对话框仍显示一个未知项.

Bug: https://pms.uniontech.com/zentao/bug-view-102148.html
Change-Id: Iacced8efe723481acd459ac7eed03c7eb68eba58