fix: cannot enter system/data/unmounted disk via "Enter/Return"
this is caused by a bug fix. while editing the name of a disk and press enter, and then entered
the inner directory, but what we need is to accept the editing.
to fix it, add a new Role named IsEditingRole, when editing delegate is generated, this data is
setted to true, and when editing finished, setted to false.
so, when enter on a disk, judge this role's data rather than judge if it is editable, which is
more accurate.

Log: fix issue: cannot access system/data/unmounted disk when press enter on it.

Bug: https://pms.uniontech.com/zentao/bug-view-106996.html
Change-Id: I2b96e4927eab506532d7cb372672e0f19144294f
(cherry picked from commit 01d6d90a649a91b10405c9dc6abaef14694b4aef)