fix: When creating a canvas object, the obtained screen data is empty, resulting in desktop jamming
During the process of creating a canvas object, the screen is deleted and the obtained screen data is empty, resulting in some logic initialization errors.

If the screen data is empty, it indicates that the current screen state has changed and the creation is abandoned.

Log: Fix the problem of desktop jamming caused by multiple quick plugging of the screen
Change-Id: I339ec8d341593024b01bd89ecb1342b39e56fe35