fix: network disconnected cause computer view shown two smb item.
the 'fake' smb item is shown if no 'real' smb item founded. when network disconnected, the 'real' smb item will not be added to the container which was used to judeg whether to show the 'fake' item. and the 'real' item is added only when it it 'existed', but the `exist` function return false when offline.
so add the 'real' to the container no matter if it is `existed` or not.

Log: fix computer page display issue

Bug: https://pms.uniontech.com/zentao/bug-view-109875.html
Change-Id: Ie4992e4c3bc61a0e036570cc17318aaa3320519c