fix: [computer] computer view is empty if default view is not computer view.
if default is not computer, the computer view is initialized when cd to computer, and the signal itemQueryFinished is emitted when window is created, on that moment, there is no receiver.
so cache the result of itemQueryFinshed, when construct computerModel, get the cached items in constructor.

Log: fix computer view shows empty if it is not default view.

Task: https://pms.uniontech.com/zentao/task-view-89648.html
Change-Id: I47eaf8db12b2e064380e10517745925a07a672bd