fix: [computer] the encrypted device's size will not update
the size of encrypted device is obtained from clear block device which is not shown, so when the clear block device's size changed, there is no receiver to handle the changed.
so we should update the encrypted backing device which is the shell of clear block device.

Log: fix issue

Task: https://pms.uniontech.com/zentao/task-view-89648.html
Change-Id: I568ce9a5c7014ef1209971389047276ba2a5b65e