fix: Failure to load third-party ext plugins
When the file manager is running, copy the third-party plugin, the text manager will load it. But when loading, the plugin may not be copied to completion, so it will fail to load. This modification adds `retry` mechanism to repeatedly try to load the plugin. (For now there is no way to know if a file is actually copied)

Log: 修复加载第三方插件失败的问题

Bug: https://pms.uniontech.com/zentao/bug-view-112007.html
Change-Id: Ibad7e1c9470e95411200344a38533cc5e4d2cd73