fix: [net share] cannot open share file in certain case
if a network share file's (smb/ftp...) name starts with '#', the url is
truncated at the pos of '#' by QUrl,  and mimetype of parent directory
is obtained, so a new fm window is opened and violate the user's
expectation.

tried construct url by percentage encoded string, but it's useless, so
if a file path contains '#', use gio to obtain the mimetype.

Log: fix a corner issue.

Bug: https://pms.uniontech.com/bug-view-116477.html
Change-Id: Ibc918e294411fafb3200d7d0424d115140b3b93d