fix: [computer] cryptsetup-ed device cannot be shown in computer
for most cases, the block device descriptor is the same as the udisks
object path's last fragment seperate by '/', but for those devices
encrypted by `cryptesetup` command, the device descriptor is not.
the device desc is 'dm-0' while the object path for this device is /org/
freedesktop/UDisks2/block_devices/dm_2d0. so, compose url by dm_2d0
cannot map to a udisks block object path. we tried to obtain the path by
split the url 'dfmroot:///dm_2d0.localdisk', and tried to resolve the
device node '/dev/dm_2d0' in udisks manager and gets nothing. so we
should splice the url by udisks' `device` property, and the value of
..../dm_2d0 is dm-0, so we get the correct url 'dfmroot:///
dm-0.localdisk', and this ensure the subsequent process goes on.

Log: fix issue.

Bug: https://pms.uniontech.com/bug-view-118033.html
Bug: https://pms.uniontech.com/bug-view-117941.html

Change-Id: Ide1dfa864a96729462547c180c8f6781a30330f5