feat: impl `EventSequence` mechanism in framework
The feature of `EventSequence` is that the events will sequence the execution of the sequence, and once one of the sequence returns true, then the execution of the sequence will stop, and if it returns false, then it will continue to traverse. If any of the sequences return true, the final call will also result in true.

Log: 框架新增事件序列机制
Change-Id: I4cabb369369f7e094d61fe28fea4a2b32f5dc804