fix: [share] cannot list the shares of server
when first access a samba server which is never been browsed before, the
list is empty.
the root path of a samba should be mounted too.
add a machinsm that every plugin can add a prehandle function to do
something before jump to the target url.
and the prehandle function is bind with url's scheme.

Log: fix issue.

Bug: https://pms.uniontech.com/bug-view-123965.html
Change-Id: I1400ba5bddbb80ed4254b34047b8718c2444bbdd