fix: [dock] cannot eject the udisks which have a encrypted partition
1. if a udisks have a encrypted parititon, then 'isRemovable' judgement
will be invalid, and makes the unmount process stop.
add judgement that if a partition is encrypted and have a
cryptobackingdevice which is not '/', the 'isRemovable' is decided by
the cryptobackingdevice. so does when invoke poweroff,
cryptobackingdevice is passed rather than the unlocked device itself.
2. fix wrong name for unlocked device;
3. fix wrong icon for unlocked device;

Log: fix issues with dock plugin.

Bug: https://pms.uniontech.com/bug-view-124033.html
Change-Id: If1eb3edcf5dfa94fd7f3691b29a84c21554de037