fix: [workspace] fetchMore function usage moidfy
fetchMore function will be invoked in various of situations, such as
whell at bottom, resize event triggered, etc.
the fetchMore function is designed to travers huge files in seperate
invokations, and every result of invoke is different from last, and is a
dynamic travers machinsm.
but we travers all the files in a single fetchMore invokation.

Log: fix repeated invokation.

Bug: https://pms.uniontech.com/bug-view-123965.html
Change-Id: Ic668738d18ec459727a49d6e751fd53dcefc8766