fix: [computer] still mount event if auto mount is disabled.
before do mount in desktop, we will ensure that if auto-mount is
enabled, in this judge function, the value is returned from memory, and
is not sync with the local file. finally found that the handler to get
settings is a new instance seperated from the global instance in the
Application, and not setted watchChange.

Log: fix issue.

Bug: https://pms.uniontech.com/bug-view-117047.html
Change-Id: I5f10fa9e3bf6c0312d05c5b011e4705efbea8b2d