fix: [share] anonymous share has no effect.
the mode of /home/$USER is 750, and refused the anonymous access of
shared folder.
(when access with anonymous, the real user who access the shared folder
is 'nobody' which belongs to 'other', so if there is no 'other' mode,
the anonymous user cannot access the shared folder.)

Log: fix issue.

Bug: https://pms.uniontech.com/bug-view-127425.html
Change-Id: I9a82b1b14fc652dba04a7c03629a211e46153568