fix: [vault]Delete a single large file without progress popup and the file management will be stuck
Cause of the problem:  When deleting, the existing cache of file information is refreshed, which leads to the need to re-obtain the information of the file being deleted. Since cryfs locks the file when deleting the file, it leads to the phenomenon that the file management is stuck, and whether it occurs or not . remove the progress box. Processing method: Cancel flushing of file cache before deletion.

Log: code and realization feature
Bug: https://pms.uniontech.com/bug-view-129773.html
Change-Id: I774346b078e9d7a4c2cbbfe410d083560239dd74