fix: [mount/optical] cannot mount a erased dvd+rw disc.
1.and raised a 'Mount error' dialog.
after erased, it should be an empty disc, but in udisks, 'opticalBlank'
field is still 'false', and then the codes run in error handle.
so if the medium is dvd+rw disc, 'blank' filed decide by the difference
of SizeTotal and SizeFree, or the opticalBlank field in udisks.
2. fix crash when ctrlN/T at appEntry;

Log: fix issue.

Bug: https://pms.uniontech.com/bug-view-127547.html
Change-Id: If91d3f90ad21a168fc0215b394b3c36dae955d0b