fix: index changed after rename.
the index data  is setted to a file pointer at createIndex.
and the data will changed after the file is renamed but the row,column and model of index do not changed.
therefore the index of selection model differ to one of  item model.

set null data to index at createIndex to remain unchanged.

Log: 修复重命名后文件索引错误

Bug: https://pms.uniontech.com/bug-view-132079.html
Influence: 桌面重命名
Change-Id: I0006cbac13829b6f2cde4bf74d876502c7f07a86