fix: [computer] sidebar is not updated when format
when format device, the display name is changed in computer view, but
the same device's display name is not changed in sidebar.
cause the name in computer view is updated realtime, but sidebar's is
updated only when 'updateRow' is invoked.
so record the first name and when name changing is dectected in
computerModel, update the name of sidebar.

Log: fix issue.

Bug: https://pms.uniontech.com/bug-view-133319.html
Change-Id: I2f119a8d5f068f2b07c992b9a53f4a57d2edfc8e