fix: [perview]Crash when previewing very large text files
Reason: The size of the string that QString can receive cannot exceed the value range of int. If it exceeds the value range of QString, an exception will be thrown. Solution: Split the string and convert it to QString so that it will not exceed the value range of int.

Log: code and realization feature
Bug: https://pms.uniontech.com/bug-view-133271.html
Change-Id: I8f2e93dd019fcb0105d938550cc0a2cf4ae26151