fix: open with-select program-add other programs cannot launch filedialog
Use QFileDialog will call to the current function, but `WindowsService::showWindow` will call to some D-Bus interfaces in desktop.
The main thread of the desktop waits for the current function to finish running if launch filedialog from desktop, this leads to deadlocks as each side waits for the other to finish.
So this modification makes `WindowsService::showWindow` execute asynchronously,without blocking the main desktop thread.

Log: fix issue that open file dialog in desktop.

Bug: https://pms.uniontech.com/bug-view-132579.html
Change-Id: I86a3ed06a475bd3c0ce51c5c1848fea6051f941b