fix: [property]The number of folders and files displayed in the multi-file properties dialog box is inconsistent with the number displayed in the status bar.
Cause: After the source file of the linked file is deleted, the current linked file in fileinfo is judged to be a non-file, resulting in a count error. Solution: judge the linked file separately, and do not use the isFile interface to judge.

Log: Fix the problem that the number of files is incorrectly calculated in the multi-file properties dialog
Bug: https://pms.uniontech.com/bug-view-136379.html
Change-Id: I4f391bc177b28b4992628e69095f90c8009175b7