fix: [property]The display problem of the properties dialog in different scenarios
Resolved an issue where extra controls were displayed in the Recycle Bin file properties dialog. 1. Add the url processing function of the device properties dialog box; 2. Add the canHidden interface in fileinfo to determine whether the file hiding function is supported; 3. Added the canTag function to the Recycle Bin and Vault to determine whether the tagging function is supported.

Log: Fix the display problem of the properties dialog in different scenarios
Bug: https://pms.uniontech.com/bug-view-136609.html
Change-Id: I178f04552457c80dbad30fa8485675fba83d5b90