fix: [vault]After opening two file manager windows, double-click the vault on the computer page, you need to close the creation interface twice.
Reason: The double-click event processing is linked multiple times, resulting in the phenomenon that one event corresponds to multiple times. Solution: Adjust the link function of the corresponding double-click event, call the location, and avoid multiple calls.

Log: Fixed the phenomenon that the creation window appeared multiple times when the vault was in multiple windows.
Bug: https://pms.uniontech.com/bug-view-137971.html
Change-Id: I6c0e38caa1d53946f92e75c97faf498aff7ec68a