fix: [workspace]View is not refreshed after setting file hide
Reason: The filter model does not refresh the view, resulting in no display effect after filtering. Solution: After obtaining the modification event of the .hidde file, reset the filter conditions for the filter model to trigger the view refresh.

Log: Fix the problem that the settings file is hidden and the view is not refreshed.
Bug: https://pms.uniontech.com/bug-view-137859.html
Change-Id: Ia2a68d15a78e8263cfbf0b72fa4bbc2f3881e3a5