fix: [vault][workspace]There are two windows, after deleting the vault in window A, create the vault again, and there are files of the deleted vault in it
Reason: The safe is deleted by executing the rm command in the process, and filemodel does not know that the files in the vault are also deleted. So the filemodel also saves the url of the vault file. After re-creating the safe and entering, the filemodel directly obtains the url from the previously saved data. Solution: When deleting the vault, filemodel will receive the event that the root directory of the vault has been deleted, judge whether it is the root directory in the event, and clear the filemodel to save the file url.

Log: When repairing multiple windows, the vault is created after deletion, and the previous files still exist in the vault.
Bug: https://pms.uniontech.com/bug-view-138473.html
Change-Id: I51a2e9b10fcfb4719f0897e667fdd7b3b35da310