fix: [vault]Copying the computer from the data disk, the recycle bin to the safe does not display the application icon and cannot be double-clicked to open.
Reason: The files copied to the vault are not filtered. Solution: For desktop special files such as computers and recycle bins, they will be filtered out during the copying process. Do not let it be copied to the vault.

Log: Fix the problem that the desktop icons of the computer and the recycle bin are copied to the vault.
Bug: https://pms.uniontech.com/bug-view-138671.html
Change-Id: Ie95e3636c2add0cc621520d0be7728638fad2547