fix: [vault]After opening multiple safe windows, the safe is locked in a new window, but some windows do not jump to the computer interface
Reason: The window ID of the safe cannot be obtained without operating the window of the safe. After the safe is locked, some windows are not redirected to the computer page. Solution: Add the event of entering the fileview view in the workspace, and record or delete the window id in the safe.

Log: Fixed the safe being locked in a new window after opening multiple safe windows, but some windows did not jump to the computer page.
Bug: https://pms.uniontech.com/bug-view-138943.html
Change-Id: Ie4fca9f18fe8a99736cd7d367d0e71b40a1f16fc