fix: [vault]The button of the file dialog window that pops up when the vault is saving the key is displayed abnormally.
Reason: DFileChooserEdit control, because when setting the dialog object to this control, there is no setting information for the dialog. Solution: Set the dialog information first. Then set the dialog to the DFileChooserEdit control.

Log: Fixed the abnormal display of the file dialog box that popped up when the vault was saving the key.
Bug: https://pms.uniontech.com/bug-view-139059.html
Change-Id: Ib513b355fdd7028e53a9004164c539dff8a442e6