fix: [fileoperation]ctrl+z exception after deleting file
Reason: After the file is deleted, the undo operation is performed, and there is a corresponding operation on the file before the deletion, and the undo exception occurs. If the file is deleted, and the file is still operated during the undo, an abnormal prompt will appear. Solution: When performing the undo operation, determine whether the file exists.

Log: Fix the problem of undo exception after deleting files
Bug: https://pms.uniontech.com/bug-view-138921.html
Change-Id: Iae279e23b3fc357375447600a7907dc310a45f03