fix: [vault]The search result page in the vault exposes the real path of the vault.
Reasons: 1. The two functions fileurl and fileinfo in the vault iterator return local file information; 2. The redirectedFileUrl function in vaultfileinfo does not implement the return information of the vault. Solution: Modify the fileurl and fileinfo functions, and implement the redirectedFileUrl function of the vault.

Log: Fixed the problem that the search page displayed the real path of the vault file.
Bug: https://pms.uniontech.com/bug-view-138701.html
Change-Id: I684b800cf630b9561d45d3dcd609319301056b4a