fix: [vault]After the vault is locked/deleted, you can drag and drop files to the vault directory in the sidebar
Reasons: Unprocessed, vault is not unlocked, drag and drop files into a scene that cannot be added in the sidebar. Solution: Judge the current security status of candrop in vaultinfo. Returns whether drag and drop of files into the vault is supported.

Log: Fix the safe is not unlocked, files can be dragged to the vault sidebar options.
Bug: https://pms.uniontech.com/bug-view-138611.html
Change-Id: I254f4dbb16fa623f5a5d7f4532e334da2e3e7f1d