fix: cannot drag file to disc
In the old logic the `isWriteble` interface of fileinfo is used to determine if dragging is possible, but the root url of disc is a mount point, which is a read-only directory, and therefore cannot be dragged. This commit uses the hook event to handle it.

Log: fix disc issue

Bug: https://pms.uniontech.com/bug-view-138281.html
Change-Id: I0cb0fe555515f736e831919a101cd2e27eab4cf9