fix: [vault]Operating the files in the vault, the file manager is very laggy.
Reason: When obtaining the status of the safe, using QStorageInfo to obtain the file system type will cause a freeze if too many times are obtained in a short period of time. Solution: Use the interface provided by dfm-io to get the file system type.

Log: Fix the problem that the operation of the vault file is stuck.
Bug: https://pms.uniontech.com/bug-view-140787.html
Change-Id: I134e6acb13513e91bd423b7940d5cd271c1cb45f