fix: [mount/sftp] cannot mount sftp.
before do mount, network connection will be checked, and using port 21
to check connection while for sftp, port 22 is used, so network check
failed with port 21 and is regard as network disconnected, the mount
process interrapted.
port 22 should be used to check network if sftp is going to mounting.

Log: fix issue.

Bug: https://pms.uniontech.com/bug-view-140879.html
Change-Id: I2db2801958a1d0fce5523294649d7b79296c1284