fix: [vault]The search result page in the vault exposes the real path of the vault.
Reason: There is no shceme in the displayed virtual path, and the method of obtaining the path is not applicable.
Solution: Add the fileDisplayPath interface to the fileinfo base class, and the vault implements its own display path return.

Log: Fixed the problem that the search page displayed the real path of the vault file.
Bug: https://pms.uniontech.com/bug-view-138701.html
Change-Id: I582ef0fe7715302f5fc40f4acd94ccd01e0439cd