fix: [detail]Select any file, then click the details button in the upper right corner, the details of the selected file are not displayed.
Reason: The window id does not match, and it is caused by the display and recycling methods. Solution: Create a detail object every time a new window is opened and use "map" to save the object and window id; recycle the detail object when closing the window and remove the corresponding id and object from "map".

Log: Fix the issue that the selected file details are not displayed.
Bug: https://pms.uniontech.com/bug-view-141557.html
Change-Id: I9cdeeb5e0588e60a93c70e597dd75e7dfbcc40bf