fix: [fileoperations][vault]Unable to rename multiple files in vault
Reason: The file url processed by the multi-file renaming is local, and the vault is a custom url, so the renaming fails. Solution: Add the corresponding function pointer in fileoperations and add the call of the function pointer, and implement your own processing in the vault and give it to fileoperations.

Log: Fixed the issue of renaming multiple files in the vault.
Bug: https://pms.uniontech.com/bug-view-142405.html
Change-Id: Ie83a788b6e6b864a3edc9d6bf8966c7e5bb81df8