fix: [fileop] cannot traverse phone's files sometimes.
1. construct dfm-io's enumerator object timeouted after default 2 secs,
it's not enough for mtp/gphoto files, via tested, the command gio list
mtp://xxxx costs more than 2 secs.
so the 'setTimeout' function is added to dfm-io and if not setted, the
default value is 2 minites, if timeout, the construction will be give
up.
and make some optimize for seperate the path of gvfs files, avoid using
QStorageInfo to get the mountpoint of gvfsd but just string compares.
2. a huge adjust of global defines, add more detail namespace

Log: fix issue and optimize

Bug: https://pms.uniontech.com/bug-view-141939.html
Change-Id: Ie2df37e3f28dbaee895d8773e939f9c1c62799a0