fix: [workspace] cannot reaccess files of udisks after device unmount.
UT cannot cover the UI logic.
after device unmount, the FileInfo gets to invalid, but the fileview
still present the deleted url, so when device mounted and reaccess it,
the root url of file view is not changed so it will not refresh the
contents.
to fix this, redirect to an always exists path after device removed.

Log: fix issue

Bug: https://pms.uniontech.com/bug-view-155335.html
Change-Id: I65beefc4803953a32a184337c554b7d613dd4b0c