fix: [dialog] passwd dialog can be hidden by main window
when mount a network share, the exec-ed password dialog can be shadowed
by mainwindow and also blocks the mainwindow, which means you cannot do
anything in this situation but only kill the process IN WAYLAND.
give a parent widget to the dialog in wayland to avoid this.
BUT if a parent is setted to dialog, then exec blocks all windows, which
is not GOOD, so only give parent in wayland.

Log: fix wayland issue.

Bug: https://pms.uniontech.com/bug-view-164597.html
Change-Id: Ibc88883bc257c2a3260f0ff1a766624986c4e663