fix: Multi tab page. After searching and filtering, switch back and no filter conditions will be displayed
When opened by multiple labels, the display status of the current filter criteria is not recorded. Modify the view of the tab to save the display status of the current filter criteria. After switching labels, use the start status to restore

Log: Modify the view of the label search criteria box
Bug: https://pms.uniontech.com/bug-view-162789.html https://pms.uniontech.com/bug-view-162791.html