chore: [mount] some cases in certain situation
if a bookmark item of smb share is added before smb daemon mount, after
daemon mount is applied, the item click can trigger the mount process
but the mountpoint is not gvfs', so convert the path when mount
finished.
an example:
/run/user/1000/gvfs/smbxxxxxx/share1 is the url that bookmark item
storaged, after daemon mount, the new mountpoint is
/media/$user/smbmounts/share1, convert the target url from /run/... to /
media/...

Log: fix issue in certain situation.
Change-Id: I334e1becf1c321f4da05105e3d2513d96abec917