fix: file name disappearance and failure to rename
Select a file with a longer file name, right-click the file and name disappears。The long file name display uses expandItem and setEditData. When you right-click, it is true to judge whether the current index is in the selected state isPersistentEditorOpen. At this time, closePersistentEditor is used to close the editing state, so long files cannot be displayed, and renaming is hidden

Log: Repair file name disappearance and failure to rename
Bug: https://pms.uniontech.com/bug-view-170931.html https://pms.uniontech.com/bug-view-172299.html
Change-Id: I82ad47788cf47b0051165175a80ecb0824b75f90