feat: Add preview view of the Dci Icon
Add control settings in the preview view according to the
unique properties of the DCI icon file. The icon size and
the device pixel ratio can control the size of the file,
the mode and the theme property can control the status of
the icon, and the palette property can control the preview
color of the icon.

In addition, To preview the effect of icons under different
backgrounds by controlling the background of the view. You
can view the details of the icon by zooming in on the icon
in the view.

Log: Add preview view of the Dci Icon

Task: https://pms.uniontech.com/task-view-216579.html
Change-Id: I04c5166f1b94dafab5f37292cb63cdb6237bd9d7