fix: After modifying the file name in ftp, open the recently used file to prompt an error
The error here is to modify the name of the file in ftp.  The file in use recently should not be realistic. Add the logic processing of filerename, modify the slot functions of filedelete and fileAttributeChanged in the recentwatcher, and determine that the scheme of the file is ftp and smb to obtain your own local file

Log: After modifying the file name in ftp, open the recently used file to prompt an error.
Bug: https://pms.uniontech.com/bug-view-181443.html
Change-Id: Iab294b04385f24ea57ed7965775fae15c25542c3