fix: Icon not restored
Delete the desktopfile file (but not the end of. desktop) from the desktop to the Recycle Bin. Restore this file to the desktop. The dfmio library used by the desktop to create fileinfo got the mimeTypeName error, so the desktopfileinfo was not created. At present, there is no dfmio library code. For temporary repair, use the method on v20 to obtain mimeType

Log: Icon not restored
Bug: https://pms.uniontech.com/bug-view-183629.html
Change-Id: Ic77f441cfbdb8cabd3b3ba92b6c9fc5a785c12e2