fix: Unable to drag back to desktop
When handling the drag event, the trash is used as the scheme to determine whether it is on the same disk. It is judged that different disks are cut. Modify the judgment here to use the source file path for the same disk judgment

Log: Fix Unable to drag back to desktop
Bug: https://pms.uniontech.com/bug-view-183395.html
Change-Id: Iea0fb3a7970bb29c6e6c137b338ca4f6e463bdd4