fix: optimize menu for new file
Only the basic functions are implemented in the menu scene. If you need to add interactive or extended functions to the menu actions, you need to filter and change the menu items in the upper module and rewrite the menu actions.

Add default processing in the NewCreateMenu, file manager and desktop filter its actions to achieve creating files and interactiving effects.

Log:

Bug: https://pms.uniontech.com/bug-view-185581.html
Influence: 右键菜单
Change-Id: Id38b2b4ea3a2054166844ef2131ea91b3afc0953