fix: Cut an empty folder and paste it to a USB flash drive slowly
When cutting to a peripheral, the size of the directory and linked file is not counted. Finally, wait for a few seconds to read the write block size on the device. Modify here to count the size of directories and linked files when cutting to peripherals

Log: fix Cut an empty folder and paste it to a USB flash drive slowly
Bug: https://pms.uniontech.com/bug-view-186879.html
Change-Id: Icde00120b2aa84615f958777d95dc77a55b32ebb