fix: There is no file display on the desktop. There is a problem with switching virtual directory display
Adjust the code iterator in dfmbase to the previous one. The civil servants use the iterator to inherit and use, and the signal does not change. Modify the map in filedatamanager to use value to access the value.

Log: There is no file display on the desktop. There is a problem with switching virtual directory display
Change-Id: I7a1d5c2dbff0b04baee3040ce184144272cda455
(cherry picked from commit 96bce3938a3db20e1034e83f0f00d88f4d3cccfb)