fix:[workspace][sort] File manager crashed
When deleting the mtp file, the scheme of the url of the file monitored by the monitor starts with mtp, and all created fileinfo is null ptr. Modify the path that is converted to a local file when the monitoring signal is received, and create fileinfo.

Log: fix File manager crashed.
Bug: https://pms.uniontech.com/bug-view-189655.html
Change-Id: Ib443581b26ef172e6db6e9d45f4c99e67e8bf556