fix: [io][copy]Completely delete the file and the desktop is deadlock
When deleting peripheral network files, the statistics thread is used to count the files. When the deletion is completed, the statistics thread has not finished counting, and the statistics thread is destructed without waiting for the statistics thread to exit. However, the statistics thread uses a QList<QUrl>to insert and reallocate the internal memory (but the location is not there, which is a wild pointer), resulting in memory problems elsewhere. Modify to wait for the end of the statistical thread in the thread, and then execute the destruction after receiving the end signal.

Log: Completely delete the file and the desktop is stuck.
Bug: https://pms.uniontech.com/bug-view-189653.html
Change-Id: Ica7a849c0e188d5817dba3f32bbe60d20b3dfd07