fix: [io][udevice][copy][sort]U disk cannot be unloaded, sorting error, file manager crash
There was an error writing the copy file to the peripheral. The file descriptor was not closed. When the sorting directory is sorted in descending order by default, the file is on the top when the device outside the disk is sorted. The monitor remove message was received when the phone was unplugged. The wrong url was removed. When modifying, the file descriptor is closed in case of an error, and the directory is kept at the front when it is not mixed, and the modification is removed by index.

Log: The USB flash disk cannot be unloaded and is sorted incorrectly。
Change-Id: Ib1c450c742c81239b3da3219c5c2483b62010e6a