fix: [io][trash]Cannot delete a folder with a long name
Adjust the delete and restore process. When deleting, save the starting time of the deletion. When restoring, restore through the recycle bin following the previous process. Use ctrl+z to retrieve the start time of the deletion, create an iterator for trash, and compare the url and the deletion time

Log: [io][trash]Cannot delete a folder with a long name
Bug: https://pms.uniontech.com/bug-view-191087.html
Change-Id: Ic12b8fce438fa8a2c06d2f0bef0d4747538d73f0