fix: Right click to create a new "Office Document" on a USB flash drive or inside an SMB, and the text icon will appear
When creating a file, create an asynchronous fileinfo, so if the suffix is empty, fill in the template error. Modify the use of the suffix interface, which is a synchronization interface in abstractfileinfo. To go to suffix, only the URL is required, and there will be no stumbling or blocking.

Log: Right click to create a new "Office Document" on a USB flash drive or inside an SMB, and the text icon will appear
Bug: https://pms.uniontech.com/bug-view-194115.html