fix: Empty Trash prompt file incorrect
Because when counting the files in the recycle bin, the count of gio's trash terms was incorrect, with long file names or other attachments to/data/home/ Caused by the path. Modify the use of asynchronous self counting of the number of trashes, and follow the subsequent process.

Log: Empty Trash prompt file incorrect
Bug: https://pms.uniontech.com/bug-view-194085.html