fix: Copying a large number of files from a mobile disk to the local location, skipping large folders
The asynchronous file information has not yet returned, so the filecount attribute was obtained as 0, skipping the current directory. Is the modification done by synchronizing file information and failing to determine if filecount is 0 (obtaining filecount means creating an iterator to obtain it), or by creating an iterator to iterate.

Log: Copying a large number of files from a mobile disk to the local location, skipping large folders
Bug: https://pms.uniontech.com/bug-view-194699.html