fix: Open the trash, deleted files are constantly refreshed multiple times
The fieldinfo in the recycle bin does not use caching, resulting in the drawing of corner markers constantly obtaining fileinfo, which is asynchronous. Every time the retrieval is completed, the interface needs to be refreshed, and the recycle bin operation file gets stuck. Modify the usage cache.

Log: Open the trash, deleted files are constantly refreshed multiple times
Bug: https://pms.uniontech.com/bug-view-195395.html