fix: After restoration, the icons on the desktop are displayed as jagged icons
This issue occurs because when deleting a file, it is incorrect to determine whether the desktopfileinfo is correct, but at this time the file's fileifno is cached. When creating a fileinfo, the local file is not cached (which does not affect any performance).

Log: After restoration, the icons on the desktop are displayed as jagged icons
Bug: https://pms.uniontech.com/bug-view-198159.html