fix: The speed of cutting files from the trash to USB drive is very slow
At the same time, cutting was performed on the USB drive and the rename operation was performed, but no files were written. The statistics progress of how many sectors were read and written to the file on the USB flash drive, so the statistics progress was used incorrectly. Write size when adding rename, count reprocessing

Log: The speed of cutting files from the trash to USB drive is very slow
Bug: https://pms.uniontech.com/bug-view-198449.html