fix: [trash/property]The icon in the right-click attribute of the trash file displays incorrectly
The fileinfo used in the recycle bin does not use caching, and the proxy fileinfo for files deleted using SMB is asynchronous, as the corresponding file attributes have not been obtained yet. Modify and add the slot function, and the fileinfo update is completed before updating the icon display.

Log: The icon in the right-click attribute of the trash file displays incorrectly
Bug: https://pms.uniontech.com/bug-view-198943.html