fix: [io/workspace]Rename unsuccessful
An asynchronous FileInfo was created before renaming, causing information to be queried for files under mtp, resulting in MTP device busy at the time of renaming. The previous MTP information query completed when the file was added on the interface, but this is the file information query that was not created the last time the file was added, so it determines whether the file exists or is an error. Modify 1. Fileinfo is created synchronously before renaming (asynchronous for next phase optimization, waiting for asynchronous queries to complete). 2. When adding a file, do the file information again

Log: Rename unsuccessful
Bug: https://pms.uniontech.com/bug-view-197913.html