fix: [taskdialog]Every time a small file is copied, the position of the copy window will be refreshed
The size of each item will be adjusted when the height of the copy progress bar is changed. Here, the position of the entire display dialog box is adjusted. Modify without adjusting the entire display dialogue position.

Log: Every time a small file is copied, the position of the copy window will be refreshed
Bug: https://pms.uniontech.com/bug-view-199201.html https://pms.uniontech.com/bug-view-199209.html