fix: [trash/restore]File restore, file manager occasionally crashes
The recycle bin is fully restored, and the iterator iterates out two files (actually, these two files are one). There was an issue with the first replacement and it was not recorded. The next file was processed directly. Modify 1. Record the URLs that have been manipulated, and those that have been manipulated are no longer being processed. 2. Modify the signal sending in the taskwidget, and I will call adjustsize when creating a new btn and adding a widget.

Log: File restore, file manager occasionally crashes
Bug: https://pms.uniontech.com/bug-view-200815.html