feat: Optimize ctrl+a, ctrl+c optimization
Modify the ctrl+a select all code. 1. Modify the flags in the fileviewmodel without using fileinfo. 2. Modify the data function in the fileitemdata class to read properties without fileinfo. 3. Modify the itemselect function in the basestatusbar to display multiple selected items without traversing fileinfo. 4. Remove the fileinfo judgment from the fileitemdata function in the fileviewmodel. 5. Remove the fileinfo judgment from the fileitemdata in the selectedIndexes function in the fileview.Modify the code copied by ctrl+c. 1. Modify the printing and use qdebug to print the selected URL. 2. Modify QClipboard:: dataChanged signal to queue connection in clipboard. 3. Modify the slot function onClipboardDataChanged in Clipboard without obtaining fileinfo. 4. Modify the use of inode to determine whether the linked file is cut.

Log: Optimize ctrl+a, ctrl+c optimization